Reduce hierarchy depth in the API folder a little
