+++ Ninja Price/API/PoeNinja/DeliriumOrb.cs
@@ -1,9 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace Ninja_Price.API.PoeNinja.Classes;
+namespace Ninja_Price.API.PoeNinja;
 
-public class Vials
+public class DeliriumOrb
 {
     public class RootObject
     {
