+++ Ninja Price/API/PoeNinja/Fossils.cs
@@ -2,9 +2,9 @@
 using System;
 using System.Collections.Generic;
 
-namespace Ninja_Price.API.PoeNinja.Classes;
+namespace Ninja_Price.API.PoeNinja;
 
-public class UniqueArmours
+public class Fossils
 {
     public class RootObject
     {
@@ -29,14 +29,14 @@ public class UniqueArmours
         [JsonProperty("levelRequired", NullValueHandling = NullValueHandling.Ignore)]
         public long? LevelRequired { get; set; }
 
-        [JsonProperty("baseType", NullValueHandling = NullValueHandling.Ignore)]
-        public string BaseType { get; set; }
+        [JsonProperty("baseType")]
+        public object BaseType { get; set; }
 
         [JsonProperty("stackSize", NullValueHandling = NullValueHandling.Ignore)]
         public long? StackSize { get; set; }
 
         [JsonProperty("variant")]
-        public string Variant { get; set; }
+        public object Variant { get; set; }
 
         [JsonProperty("prophecyText")]
         public object ProphecyText { get; set; }
@@ -57,10 +57,10 @@ public class UniqueArmours
         public Sparkline LowConfidenceSparkline { get; set; }
 
         [JsonProperty("implicitModifiers", NullValueHandling = NullValueHandling.Ignore)]
-        public List<PlicitModifier> ImplicitModifiers { get; set; }
+        public List<object> ImplicitModifiers { get; set; }
 
         [JsonProperty("explicitModifiers", NullValueHandling = NullValueHandling.Ignore)]
-        public List<PlicitModifier> ExplicitModifiers { get; set; }
+        public List<ExplicitModifier> ExplicitModifiers { get; set; }
 
         [JsonProperty("flavourText", NullValueHandling = NullValueHandling.Ignore)]
         public string FlavourText { get; set; }
@@ -88,7 +88,7 @@ public class UniqueArmours
         public string DetailsId { get; set; }
     }
 
-    public class PlicitModifier
+    public class ExplicitModifier
     {
         [JsonProperty("text", NullValueHandling = NullValueHandling.Ignore)]
         public string Text { get; set; }
@@ -105,5 +105,6 @@ public class UniqueArmours
         [JsonProperty("totalChange", NullValueHandling = NullValueHandling.Ignore)]
         public double? TotalChange { get; set; }
     }
+
 
 }
