+++ Ninja Price/API/PoeNinja/Leagues.cs
