+++ Ninja Price/API/PoeNinja/Vials.cs
@@ -1,9 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace Ninja_Price.API.PoeNinja.Classes;
+namespace Ninja_Price.API.PoeNinja;
 
-public class DeliriumOrb
+public class Vials
 {
     public class RootObject
     {
